backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
pymongo>=4.15.4
python-dotenv>=1.1.1
openai>=1.0.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
werkzeug>=3.0.0
pytest>=7.4.0